requirement.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
Django==3.2.22
django-cors-headers==4.2.0
django-filter==22.1
django-redis==5.4.0
django-simpleui==2023.8.28
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
docker==6.0.1
idna==3.4
importlib-metadata==6.1.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
Markdown==3.4.2
MarkupSafe==2.1.2
packaging==23.1
paho-mqtt==1.6.1
Pillow==9.4.0
prompt-toolkit==3.0.39
psutil==5.9.4
pycparser==2.21
pycryptodomex==3.19.0
Pygments==2.14.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==23.2.0
python-alipay-sdk==3.3.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==5.0.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tencentcloud-sdk-python==3.0.998
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.8
websocket-client==1.5.1
zipp==3.15.0
